Reorganize tagged.rs module
