+++ src-tauri/src/units/tagged/base.rs
@@ -1,9 +1,8 @@
 
-use super::composite::CompositeUnit;
-use super::convertible::UnitConvertible;
-use crate::util::prism::ErrorWithPayload;
+use crate::units::composite::CompositeUnit;
+use crate::units::convertible::UnitConvertible;
+use super::error::TryConvertError;
 
-use thiserror::Error;
 use num::One;
 
 use std::fmt::{self, Formatter, Display, Debug};
@@ -13,13 +12,6 @@ use std::fmt::{self, Formatter, Display, Debug};
 pub struct Tagged<S, U> {
   pub value: S,
   pub unit: CompositeUnit<U>,
-}
-
-#[derive(Clone, Debug, Error)]
-#[error("Failed to convert units")]
-pub struct TryConvertError<S, U> {
-  pub tagged_value: Tagged<S, U>,
-  pub attempted_target: CompositeUnit<U>,
 }
 
 impl<S, U> Tagged<S, U> {
@@ -67,11 +59,5 @@ impl<S: Display, U> Display for Tagged<S, U> {
     } else {
       write!(f, "{} {}", self.value, self.unit)
     }
-  }
-}
-
-impl<S: Debug, U: Debug> ErrorWithPayload<Tagged<S, U>> for TryConvertError<S, U> {
-  fn recover_payload(self) -> Tagged<S, U> {
-    self.tagged_value
   }
 }
