+++ src-tauri/src/units/tagged/mod.rs
@@ -1,0 +1,6 @@
+
+mod base;
+mod error;
+
+pub use base::Tagged;
+pub use error::TryConvertError;
